Ajout d'un tableau des produit sur la vue restock Plus divers changement mineurs

+++ src/caisse/product/PurchasedProduct.java
@@ -23,7 +23,7 @@ public class PurchasedProduct {
 		this.name = name;
 	}
 
-	public double getPurchasePrice() {
+	public int getPurchasePrice() {
 		return purchasePrice;
 	}
